make equivalence testing helpers more generic
